tec: Extract methods related to email addresses
